Include cdbexplain.c in explain.c
Since the functions in cdbexplan.c will need to use the new static
API in explain.c, rather than exposing that let's just #include
the file as it's only used from explain.c. The alternative would
be to concatenate the contents into explain.c but that wouldn't
improve mergeability and readability (since the result would be
huge) so let's take this opportunity to move GPDB specific code
out from explain.c instead.

This is not the end result as the file should be moved as well
to src/backends/commands, but that will be a later follow-up
commit.