Convert settings output to new EXPLAIN format
Rather than returning a formatted list, return a List of GUCs such
that the EXPLAIN code can choose how to print it. Also move the GUC
output in under verbose in the process, and make printing the current
optimizer a setting as well.

This also updates the only other consumer of gp_guc_list_show() in
allpaths.c