Refactor to use appendStringInfoSpaces rather than Fill
Since all uses of appendStringInfoFill() in the core code was adding
spaces, remove the function and revert to using the upstream version
appendStringInfoSpaces().

The two callsites that append nulls rather than spaces are both in
extensions, so copy the code there and leave them to maintain the
functionality (they both have duplicated StringInfo extensions in
the code already so another function won't break the camels back).