Another attempt at fixing auth_constraints
Since we no longer can trigger on DML against the system catalog,
add the superuser check WARNING that was lost from the trigger to
check_auth_time_constraints().

In order for a time constraint to be placed on a superuser, another
superuser must override the system catalog DML protection and
insert it directly into the catalog bypassing the normal ALTER ROLE
syntax. The warning the trigger gave won't deter an admin that will
go to such lengths, so for now lets keep it simple and add a check
such that it can easily be picked up at least.