Accept plan change in 'eagerfree' test.
This join seems totally symmetric, with two rows on both sides. It's more
or less random, which way 'round the join is performed.

I'm not sure what this query is supposed to test. Eager free of something,
based on the name of the file, but I can't tell exactly. In any case, I'm
fairly sure the new plan works just as well for that.