Cherry-pick join removal bug fixes.
I took the contents of analyzejoins.c from the tip of REL9_0_STABLE of
analyzejoins.c, and cherry-picked commit the rest of changes from upstream
commit bab9f1d585b.

This was needed to fix a few failures where the planner failed to find any
legal joins. This showed up as  errors like this in the 'information_schema'
test:

ERROR:  Query requires a feature that has been disabled by a configuration setting.
DETAIL:  Could not devise a query plan for the given query.

There are more errors with that message in other tests that this doesn't
seem to fix yet, though. I'll keep investigating those, but this is clearly
a step in the right direction.