Silence failure by fixing expected output.
The differences seemed to be because of the issue I reported in:
https://github.com/greenplum-db/gpdb/issues/4488

Now these queries are producing the same result as in PostgreSQL, which I
think is correct. Why did these queries get fixed in the merge? I don't know.
The issue in general has not been fixed, the reduced test case I reported
in the github issue still returns the same result as on GPDB 5. But this
isn't a new issue, and we surely don't need to be bug-compatbile with GPDB5,
so now that these queries produce the correct result, let's just be happy
and move on.