Silence 'update_gp' failure, by memorizing new plans.
The difference in both of these plans was that an InitPlan was at a slightly
different place in the tree. I think that's because with the merge, the
top node in the plan tree of an INSERT/UPDATE/DELETE is a ModifyTable, so
the InitPlan can now be attached to that. Before, the InitPlan had to be
attached to whatever the topmost node was (which is now the node just below
the ModifyTable).