Silence 'gp_explain' failure, by adapting it to the new plans.
Daniel did this earlier already, in commit a6256ba196, but reverted it
after I complained that we don't really expect the plan to change.

I've changed my mind on that. I still don't think we should be switching
to nested loop joins here. But we need to move on, and:

1. This test case isn't really about testing plan stability. It's about
   testing the EXPLAIN output, and the new plan is just as good for that
   purpose.

2. The way we "fall back", if a plan cannot be found with some join types
   disabled, is a bit iffy. If we e.g. need to use a nested loop join
   once in the plan, then the planner will not hesitate to use it again,
   later in the planning. That's what seems to happen in this case. We
   tend to do that more, after this merge. Which is unfortunate. But I'm
   hoping to get https://github.com/greenplum-db/gpdb/pull/4477 pushed
   fairly soon, which changes the way all that works.