Silence failure with ORCA. Add a FIXME.
This is a bug we need to fix. But it seems like a pre-existing issue, it
was only revealed now because we added btree and hash operator classes to
hstore, and tests to exercise them. I bet you could reproduce the error
with any other user-defined datatype with btree and hash opclasses.