Fix 'storage' test failure.
The proximate cause was that now that the flat files are gone, the XID
warn and stop limits, in shared memory, are only updated when VACUUM runs.
Previously, they were updated whenever the flat file was rebuilt, which
included VACUUM, but also at database startup, and at CREATE DATABASE.

The test assumed that the limits in shared memory were set when the test
starts, which would not necessarily be true for a freshly started up
cluster. Also, after running CREATE DATABASE, it assumed that you would
immediately start to see wraparound warnings, if the new database's
datfrozenxid was old, but now that only starts to happen after the next
VACUUM updates the shared memory variables.

Also, the error messages had changed in the upstream. I restored the
hint on reaching the stop limit, to suggest lowerign xid_stop_limit GUC
again. Since it was a delibarate change in GPDB, to have that hint, let's
keep it. (Whether that suggestion with xid_stop_limit is a good idea or
not is a different story.)

There are a lot of questionable things here. Like never vacuuming
template0, and allowing its datfrozenxid to wrap around in the first place.
But I'm not going to start fixing that as part of the merge. These changes
make the regression test pass again.