Fix psql \d display of relstorage and typed tables.
The column numbers got mixed up, and we printed the relstorage field
as the "type of" table for every relation.

NOTE: This is wrong in master and 5X_STABLE as well! However, because
the query has version checks, you won't notice, unless you connect to
a 9.0-based server. We should still fix it, because it's quite possible
that someone will use an old psql to connect to a newer server.