Need to clear QueryFinishPending between every query.
We used to do it in prepare_for_client_read() and client_read_ended(),
along with the upstream QueryCancelPending flag. The clearing of
QueryCancelPending was removed from those functions in the upstream, and
we removed QueryFinishPending as well as part of the merge. But we need
to still clear it somewhere.

Fixes the 'query_finish_pending' test to not get stuck forever.