Fix 'constraints' regress test conflicts
We replaced a number of postgres tests that were incompatible with
Greenplum due to the fact that we cannot support unique constraints on
any tuple that is not a distribution key. These changes were reverted
to the original postgres version during the merge to 9.0. Because
there was no indication in the tests themselves as to why they were
changed, it was difficult to figure out which version of the tests
should be kept. Reverted back to the Greenplum version of the tests,
and added comments about why they were changed in case they get
reverted to the Postgres version again in the future.