Fix subselect tests for planner
The upstream version of the answer files for correlated subqueries
with skip-level correlations were re-added in the merge. This type of
query was not supported in postgres versions prior to 8.4, and thus
was never supported in the planner. After 8.4 versions, the planner
works, but the plan it creates is not currently parallel safe.

Until the planner is updated to create a parallel safe version of the
plan, we must continue to throw errors when planning this type of
query.