Get constraints passing again
First, put back the constraints_optimizer.source file that I
accidentally removed previously. Additionally, after
2c1152b6d850b67733946a68d918421a65a8751 was committed, a number of
error messages were changed to a pretty unhelpful error message, but
because of the way that deferred triggers work with two phase commit,
we will have to live with these messages for the time being.