Fix ALTER TABLE SET DISTRIBUTED BY
During the merge, the CommandCounterIncrement() towards the end of the
call of swap_relation_files() was removed. This was needed because the
caller in this case will try to update the options on the same
tuple. Instead of putting the CommandCounterIncrement() back at the
bottom of swap_relation_files(), we can just call it afterwards before
trying to update the options to prevent this merge issue from
happening again.