Fix analyze on ao tables with toast tables
The RangeVar node type had the 'catalogname' field added to it in the
merge, and serialization was failing when running analyze on ao
tables. Additionally, 'e2631c' refactored vacuum, but as a result, the
'relation' feild attached to a VacuumStmt was being freed after
committing a vacuum transaction, but before it was needed to vacuum
external toast tables for AO relations. This has been fixed by running
the vacuumStatement_AssignRelation function inside the vac_context.