Fix vacuum full for AO tables
How we dispatch vacuum has been refactored quite a bit. For heap
tables, this seems to work fine. For AO tables, it gets pretty
complicated with all the phases it takes. During the refactor, the AO
vacuum cleanup phase was changed to dispatch the cleanup phase for the
AO table to the segments and then dispatch the vacuum for the
auxiliary tables separately in their own transactions. This seems to
be an okay idea right now but the implementation was broken due to
multiple reasons: preassigned-oids were not being populated before
dispatch, vacuum was being skipped for AO auxiliary tables, and
primaries were trying to vacuum the AO auxiliary tables on their own
instead of waiting for the dispatch. This should be fixed now but we
should review and possibly refactor AO vacuum sometime soon.

Reference to VACUUM refactor:
https://github.com/greenplum-db/gpdb-postgres-merge/commit/e26231c909e2d93492a1c1677e8124717df5ac36

Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Taylor Vesely <tvesely@pivotal.io>