Fix VACUUM for toast tables
The primaries try to vacuum their toast tables but do not have the
pre-assigned OIDs for that because the master had not vacuumed its
toast table prior to dispatching the vacuum to the heap table. A
separate dispatch for vacuuming the toast table in its own transaction
seems to be the new thing to do here.

This is very similar to how VACUUM for AO was fixed:
https://github.com/greenplum-db/gpdb-postgres-merge/commit/125454ef110f28fef47b6c30a6d2a7f377206a03