Fix AO vacuum not vacuuming auxiliary tables
After the fix for vacuuming TOAST tables was merged, AO vacuum was not
vacuuming the auxiliary tables. This was because the vacstmt being
reused was marked as AOVAC_NONE but the auxiliary table vacuuming
condition required AOVAC_CLEANUP. To fix this mistake, we just create
new VacuumStmt objects instead of reusing the current one.

Reference for the fix on vacuuming TOAST tables:
https://github.com/greenplum-db/gpdb-postgres-merge/commit/02919309ff913606fb1d233737b9df3abcee9c49