Allow using a merge join for a dummy FULL JOIN.
Like in the 'join' regression test:

postgres=# select * from int4_tbl a full join int4_tbl b on true;
ERROR:  Query requires a feature that has been disabled by a configuration setting.
DETAIL:  Could not devise a query plan for the given query.
HINT:  Current settings:  optimizer=off

NOTE: This should be submitted as a separate PR, as it's a pre-existing
issue. We just got the test case in this merge.