Fix vacuum full heap regress expected output
Updated the expected output to match the new clustering
method. Validated by using gp_filedump on the relation and index files
to make sure things are correct.

There is still a failure in vacuum_full_heap where we expect 1 relpage
but the new cluster gives 0 relpage. This is because of a
Greenplum-specific mechanism put in place long ago where the planner
would dispatch to the segments with a request to run pg_relation_size
if the relpages for the relation in pg_class was 0. This is
technically correct but had a bad performance issue with very large
partition tables. Because of that, we mark it as 1 instead of 0 in
lazy vacuum. We do not do this in vacuum full's clustering but we
should keep the expected process. This will be added back sometime
soon which then the test will turn green.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Taylor Vesely <tvesely@pivotal.io>