Disable two-phase commit for REINDEX.
Because relmapped relations do not support PREPARE, reindexing
relmapped relations such as 'pg_class' fail with two-phase
commit. Reindex should not make any logical changes to table data, and
therefore should not be a problem if it fails on one or more
segments. Two-phase commit was initially turned on with REINDEX
because it was a requirement of the database at that time to keep the
relfilenode in sync on all segments. This is no longer a requirement
of Greenplum, so we are now able disable two-phase commit for the
REINDEX operation.

Co-authored-by: Heikki Linnakangas <hlinnakangas@pivotal.io>