VACUUM FULL now clears relpages/reltuples, adjust expected output.
For better or worse, the new VACUUM FULL implementation resets relpages
and reltuples to 0. PostgreSQL master still does that, too. I'm not sure
why we don't try a bit harder to collect correct values for these, while
we rewrite the table, but let's not try to second guess that as part of
the merge.