Quick & dirty fix for failing 'fsync' tests on concourse pipelines.
On my laptop, this test always fsync's 5 files, just like on master
pipeline. But for some reason, on the gpdb_merge pipeline, it fsyncs 6
files. I have no clue why, but the test only cares that it's at least 2
files, so I suppose that's OK. Add an alternative expected output file
for that. I couldn't figure out a better way to mask out that difference.

I hope and believe that this test can be removed altogether once we catch
up to PostgreSQL 9.2, because starting with 9.2, the checkpointer is
separate from bgwriter in PostgreSQL, too, and there shouldn't be much
GPDB-specific code left to test.