Lower xid_stop_limit, to not go over the limit after pg_resetxlog.
pg_resetxlog resets the "oldest database's datfrozenxid" value to next
XID - 2000000000. With the old default of xid_stop_limit=1000000000, you
immediately go over that, when you run pg_resetxlog. pg_upgrade was failing
because of that.

NOTE: The "stop limit" is 1000000 in PostgreSQL. I don't understand why we
have bumped it up to 1000000000 in GPDB, so I didn't dare to bring it all
the way down to the upstream value. I don't see any reason we couldn't, and
I think we should do that, but not as part of this merge.