Try another fix for the XID wraparound warnings after pg_resetxlog.
You still got warnings, the limit was still too close. Instead of changing
the defaults again, revert the change to xid_warn_limit, but use a smaller
value just in pg_upgrade, to silence the warnings there.

NOTE: We need to have a discussion on gpdb-dev on why these limits differ
from upstream.