Allow walsenders to use superuser reserved connection slots
In Greenplum, there is a concept of restricted mode where
superuser_reserved_connections is set equal to max_connections making
it so only superusers can connect. Changes made in restricted mode
need to be replicated to the standby master. We currently only support
one walsender anyways so we should allow the connection to happen.

As a note, Greenplum restricted mode is used in a couple utilities
(i.e. gpexpand) and for strict database maintenance.