WIP: Test if QueryFinish signals are being sent to wrong backend.
This is a quick & dirty patch to add the current "gp_command_count" value
to query-finish signals, to avoid signaling wrong query. Hopefully that
fixes the weird flakiness in the pipeline, where some rows seem to be
mysteriously missing from result sets.

NOTE: This is really just a quick & dirty test. This should be reverted
and replaced with a more thought-out patch, if this helps.