Remove autovacuum tests from test_xidlimits tinc tests
The merge with Postgres 9.0 changed the behavior of the clog cleanup,
so now we remove clog files that are associated with transactions from
template0. The premise of these tests was to confirm that if a table
was created from one of these aged templates, that we would be able to
sucessfully vacuum the table to bring it up to date. Now that we are
about to re-enable autovacuum for tables with datallowconn='f', this
no longer makes sense to test.