Heavily refactor copy.c, to follow the upstream's code flow more closely.
This rewrites much of the COPY FROM  dispatching code. The on-wire format
between QD and QE is now different, from starters. This is completely broken
now, but it's hard to continue without being able to test it. So I'm
committing this in its current state, and will continue cleaning this up
once we get initdb to run.

Single-row error handling is surely also broken...