Fix assertion failure, where a Var was processed twice.
There is now an assertion that setrefs.c doesn't process a Var node
more than once, i.e. that a Var's varno is not already INNER or OUTER.
We were tripping the assertion in many queries. The assertion is new,
which is why we didn't trip that before.

The fix is to revert this code to the way it is in the upstream. It was
changed in GPDB a long time ago, but I couldn't figure out why. If there
was a reason, I guess we'll find out later..