Make error handling in seqserver more appropriate.
I was getting a PANIC because the error recursion limit was reached.
I'm not yet sure what the root error was that started the recursion, but
this makes the error handling more robust, if an error happens, for any
reason. Seems like a good thing to do anyway.

NOTE: This should perhaps be done as a separate PR. The error handling
seems highly suspect to me even after this - shouldn't we send the error
to the caller or at least log it? But at least we will no longer crash..