Shore up the error handling in seqserver more.
Make it log any errors to the log.

NOTE: This probably should be done as a separate PR, as it's not really
related to the merge. It's just that the merge happened to introduce a bug
that throws an error in seqserver, and it was difficult to debug without
this.