Fix tuplesort's new 'cluster' variant.
When I merged this, I thought we'd need to do something more complicated
here, because tuplesort.c deals with MemTuples rather than HeapTuples now.
But in fact, tuplesort.c already deals with many different kinds of things,
MemTuples, IndexTuples, or plain Datum, depending on which tuplesort_begin_*
function you call. So actually, we can let the tuplesort_begin_cluster()
variant use HeapTuples, just like in the upstream, and we don't need to
do anything special. Hence, just use the upstream code for
tuplesort_getheaptuple(), and it works fine.

I haven't implemented the new cluster variant for the mksort yet, so just
fall back to the upstream tuplesort for that, to avoid the "not implemented"
elogs. This allows us to punt on that for longer.