Use "session ID" instead of "backend ID" in prefix of temp table's files
Since commit debcec7d in PostgreSQL 9.1, the current backend's backend ID
has been included as part of the filename of temp relations. However,
in GPDB, all QE processes have a different backend ID, even when they
belong to the same session, and therefore share the same temp relations.
Use "session ID" instead. We were already using the "session id" as part
of the temp schema's name, so this is consistent with that.