Refactor the way ResultRelInfos for partitions work.
We used to expand the estate->es_result_relation_info array on-the-fly,
whenever a partition of the INSERT/UPDATE/DELETE target was modified.
However, ModifyTable node keeps pointers into the es_result_relation_info
array, and trying to repalloc() invalidated those pointers, leading to
crashes. Refactor things so that the ResultRelInfos for partitions
are kept as part of the parent ResultRelInfo.