Add missing SetRemoteDestReceiverParams() call.
PostgreSQL 8.4 already refactored CreateDestReceiver like this.
The 'portal' used to be a second argument to CreateDestReceiver, but now
you need a separate call to set it. The corresponding calls in
exec_simple_query and exec_execute_message do this, and before the 8.4
merge, we passed 'portal' as the argument here, too.

This was causing a crash in the regression tests now. I don't know why
it hasn't caused trouble before, AFAICS it's been broken since the 8.4
merge.

NOTE: This ought to be done as a separate PR, since it's an 8.4 merge
issue.