Fix ALTER TABLE ADD PARTITION.
It was crashing, because CreateStmtContext.pstate was not filled in
correctly. In the passing, refactor the atpxPart_validate_spec() function
a bit. All the callers were just passing an all-zeros CreateStmtContext
as the argument, so might as well remove the argument and initialize the
dummy CreateStmtContext in the function itself.

NOTE: This refactoring would make sense on master, independently of the
merge, so would be nice to create a separate PR of this.