Fix EXPLAIN of Aggregate's outputs.
It was broken in the case that the Aggregate's output didn't include all
of the Grouping Keys. Such a plan was generated in one of the queries
in 'table_functions' test. The confusion was that the 'subplanColIdx'
refers to the *subplan's* target list, not the target list of the Agg
node itself.

NOTE: I think this has been broken all along. Separate PR?