Fix assertion failure in planning GROUP BY with two identical expressions.
If the subplan's target list contained two identical copies of the same
expression, but with two different ressortgrouprefs tlist_member() would
always find the first one. That failed the assertion that checked that
the ressortgroupref matches.

Since we really want to find the TLE with the correct ressortgroupref,
let's use get_sortgroupref_tle() instead.

I don't know why this only started failing now. Maybe we just happened to
not have a test case for it, or the planner created different targetlists
for such queries before.