COPY refactoring continued.
The most notable change is that we no longer construct a new COPY command
from scratch, when a COPY is dispatched from QD to QE. We send the original
CopyStmt, with the few additional fields filled in. This makes the
dispatchment of COPY a lot more similar to other utility commands.