Remove PITA test.
This test keeps breaking whenever any new columns are added or removed from
tables with more than 20 columns. That's a pain, we've had to change the
expected output of this test practically every time we merge. Now that I
look at this more closely, I don't see the point of this query, so let's
just remove it.

NOTE: This should be done as a separate PR.