Fix NEWLINE option handling.
NOTE: This changes behaviour! With this change, if you specify NEWLINE 'nl',
and contain a CR in the input data, you will get an error. Previously, it
would have been treated as a regular character. Now you must escape it.
This matches the upstream behaviour, and I doubt this was a well-thought
out change in GPDB, probably just a side-effect of a different
implementation.