Fix planner bug with NestLoop, IndexScan, Motion combo.
When create_motion_plan() called create_subplan(), it reset the current
"outer params" state. It's quite right that the subplan needs to be
processed indepenedetly, as you can't pass params through a Motion. But
it needs to restore the state, when it's done processing the subplan.

The query and plan that hit this looks like this:

postgres=# explain select a.t from bm_test a, bm_test b where a.n < b.n;
                                        QUERY PLAN
-------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice2; segments: 3)  (cost=0.00..2.11 rows=4 width=8)
   ->  Nested Loop  (cost=0.00..2.11 rows=2 width=8)
         Join Filter: a.n < b.n
         ->  Broadcast Motion 3:3  (slice1; segments: 3)  (cost=0.00..1.05 rows=1 width=7)
               ->  Seq Scan on bm_test b  (cost=0.00..1.01 rows=1 width=7)
         ->  Materialize  (cost=0.00..1.01 rows=1 width=15)
               ->  Seq Scan on bm_test a  (cost=0.00..1.01 rows=1 width=15)
 Optimizer: legacy query optimizer
(8 rows)

NOTE: Is this a pre-existing bug? How did this ever work?