Revert the way ERRCODE_* are matched in libpq to the upstream way.
Including errcodes.h (indirectly via elog.h) created a compile time
dependency, because errcodes.h is not generated from errcodes.txt with a
perl script. We could expose that dependency in the Makefile, like we've
done e.g. in src/port/Makefile, but seems better to just avoid the problem
in the first place. Let's just revert these macros the way they're in the
upstream.

This hopefully fixes the failure we've seen in the ubuntu concourse job.