Re-enable MIN/MAX optimization.
I'm not sure why it's been disabled. It seems trivial to make it work, so
let's to it. Might not be a very common query type, but if you happen to
have a query where it helps, it helps a lot.

NOTE: This should be done as a separate PR.