Revert MyTempSessionId() patch, stop including session_id in temp table names.
There was a fatal flaw with the idea of using session_id in temp table names.
In GPDB, we keep temporary tables in the shared buffer cache, which means
that any backend needs to be able to flush them to disk. But we don't store
the backend/session ID in the buffer descriptor, therefore a backend that
just looks at a Buffer can't tell which backend it belongs to. We could
add backend ID to the buffer descriptor, but let's try a different approach:

With this patch, we no longer include the session/backend ID in the temp
table's filename. With that change, a backend doesn't need to know which
backend a temp relation belongs to, in order to flush it to disk. Instead
of using MyTempSessionId(), we now just always mark temporary relations
with a constant TempRelBackendId value.