Fix planning of nested loop joins.
A lot of queries with nested loop join plans were erroring out:

ERROR:  node is not an index attribute

The culprit was that we were setting the root->curOuterRels state
incorrectly, when recursing into the children of a nested loop join.
We recurse into the inner plan/path first, whereas PostgreSQL recurses
into the outer side first. (We do it differently because
inject_partition_selectors_for_join() wants a Plan for inner side but
just a Path for the outer side.)