Remove negative test case.
This now "works", because 9.1 added a new concat() function with "VARIADIC
any" argument. That's not the intention of the test, though. I don't find
this test very interesting, so just remove it. We have caught up to the
upstream variadic support now, so I don't feel the need to have a lot of
extra tests on it, over what upstream has.