A more-working fix for the issue with LASJ_NOTIN and merge join paths.
Somewhat confusingly, 'mergejoin_allowed' actually means "merge joins *and*
nested loop joins are allowed". I don't want to rename it, since it's like
that in upstream too. Instead, check for JOIN_LASJ_NOTIN at the same place
where we used to. Apparently this check was simply lost in the merge.