Don't get fooled into thinking that a ModifyTable is a dynamic scan.
ModifyTable got added to the list of Plan types in explain.c, which are
passed to isDynamicScan(). It's not a Scan at all, so isDynamicScan()
got confused and returned 'true' incorrectly. This led to bogys dynamic
scan id being printed in EXPLAIN output.

To fix, make isDynamicScan() more robust, so that it will return 'false'
for any Plan node type that is not a dynamic scan.