Refactor creation of PartitionSelectors to use a Path node.
The old way, where inject_partition_selectors_for_join() was called
with a pre-created Plan for the inner side, but no Plan for the outer
side, was a bit weird. It was extra inconvenient for the caller, because
in the upstream, we always build the outer Plan first, so we had to
reverse that in GPDB. However, with 9.1, and the changes to the way
Nested Loop params are passed from the outer side to the inner side,
we must now construct the Plan for the outer side first.

To fix, refactor the inject_partition_selectors_for_join() function so
that it doesn't require a pre-created Plan for the inner side. It now
works purely on the Path representation.

NOTE: This is another good candidate for submitting as a separate PR.