Don't malloc 'compresstype' and 'orientation' strings separately.
Dealing with StdRdOptions is a bit tricky, as it needs to be malloc'd.
The GUC check/assig mechanism requires that. Avoid using separately
malloc'd strings there, to keep it a bit simpler.

This fixes occasional crashes related to setting gp_default_storage_options
in the regression tests ('dsp' test?). There are still some failures
related to the storage options GUC and reloptions, so more work remains..