Avoid holding an extra ShareUpdateExclusiveLock on toast table.
This probably doesn't matter in practice, but at least it silences the
'partition_locking' test failure.

NOTE: In 9.4, this code is refactored in the upstream, and I think the
new upstream code will have the same effect: On a newly created table,
we hold only AccessExclusiveLock, not both AccessExclusiveLock and
ShareUpdateExclusiveLock. I considered backporting that change, but went
with this smaller patch instead.