Refactor parserOpenTable, for safety and to match upstream better.
It's not very safe to call isFutureCTE() in PG_CATCH block. I saw a crash
there, because of some uninitialized/clobbered memory, while running
the 'extprotocol' tests. There were other issues there that caused the
original failure, but that led to a missing table and a failure in this
lookup. I'm not sure why we haven't seen that before, but this is a cleaner
way to write this in any case.

NOTE: This isn't really merge-related, so -> separate PR!