Fix CREATE EXTERNAL TABLE to not crash on error.
This was the true cause of the crash I saw in the gpextprotocol tests, not
the thing I tried to fix with the previous commit. The previous commit still
makes sense from a code clarity point of view, but it wasn't unsafe as I
thought when I wrote the commit message.