Don't dispatch index creations too eagerly in ALTER TABLE.
In general, an ALTER TABLE command is dispatched as one command, after
processing on the master. However, indexes for constraints were handled
differently: their IndexStmt were dispatched immediately, when the index
was created in the master. That caused errors like this:

alter table atacc1 alter column test type integer using 0;
ERROR:  relation "atacc1_test_key" already exists

That command consists of dropping the old index, and creating a new one,
but because of the early dispatch, the creation was dispatched before the
old index was dropped. I'm not entirely sure why this wasn't a problem
before the 9.1 merge, but there's been some changes around the ATPrepCmd,
so I'm not surprised. This seems less surprising in any case.