Fix code to choose name for index that backs a constraint in a partition.
I'm not sure why this stopped working with the merge, but let's let the
normal function to choose the index name do its thing. We only need this
custom GPDB-specific code, when the index name is given explicitly, because
multiple partitions can't have an index with the same name.

Fixes test failure in the 'partition' test that was caused by the index
name being different than before.