COPY FROM: work around double-cdbCopyEnd and use of freed struct
If cdbCopyEndAndFetchRejectNum disconnects the gang and throws an ERROR,
that error is caught up above in CopyFrom(), which calls
cdbCopyEndAndFetchRejectNum() a second time. Since we've already freed
the primary_writer, bad things happen:

    ERROR: invalid memory alloc request size 8556248572 (context 'COPY')

Work around this issue by bailing out early during the second call --
but it seems like we should not be making the second call in the first
place. That will take more refactoring.