BeginCopyFrom: fix stack smashing
The array we read into wasn't big enough to hold the QDtoQESignature.
Fix this, and remove all hardcoded references to the size of the
signature to (hopefully) avoid problems like this in the future.

The BinarySignature, which this code is based on, also hardcodes a
length -- but that's from upstream, so we haven't touched it here.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>