UGLY hack to fix functional deps during parallel group planning
Shameful.

With 9.1 it is now possible to select columns in a grouping query that
aren't in the GROUP BY clause, as long as those columns are from a table
that is being grouped by its primary key. However, parts of the parallel
planning code assume that all variables are part of either a GROUP BY
clause or an aggregate, and so they lose these additional columns,
leading to errors like

    ERROR:  invalid attrno 3, should be in the range [1, 2] (cdbgroup.c:4089)

To work around this, replay the discovery of functional dependencies
between variables in the parallel grouping planner. If a column comes
from a table that is being grouped by its primary key, but that column
isn't part of the GROUP BY clause, add it to the GROUP BY target list
anyway to avoid changing other assumptions in the code.

Eventually we should get rid of this magic knowledge that the parallel
grouping planner has about the plan, and consolidate the two paths.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>