Fix dispatch test failure by reconnecting to database
According to newly introduced changes to GUC framework in 9.1, if a
GUC check hook fails, the GUC's value is not changed.  The dispatch
test sets search_path of a database to include a schema that doesn't
exist.  New connections to the datbase now don't have "my schema" in
the search_path.  Prior to 9.1 changes (as it stands on master), the
search_path would include "my schema" in spite of the fact that in
doesn't exist.  This change fixes the test to reconnect to the
database after "my schema" is created, thereby setting search_path as
expected.