Handle type check error in test case with merge join.
In test case 'bfv_statistic', pg_statistics.stavalues
is updated to a different type and a type check issue
will be encountered when evaluating merge join.

In GPDB, with enable_mergejoin=off, merge join will
not be considered during planning, so this issue did not
arise.

In PG, with enable_mergejoin=off, merge join will still
be evaluated, with a huge constant cost.