Fix a couple of diffs in alter_table ICW test
Foreign key constraints are ignored with a warning in Greenplum.
Follow the same behavior for newly introduced ALTER TABLE VALIDATE
CONSTRAINT command.

The auto-generated index names during ALTER TABLE ADD CONSTRAINT
differ in Greenplum as compared with upstream.  For now, use the
Greenplum specific names in answer files but at some point, we need to
reconcile the naming scheme with upstream.

Remove a merge fixme on the way.