Fix the icw returning case failure.   -ERROR:  cannot update a view   -HINT:  You need an unconditional ON UPDATE DO INSTEAD rule.   +ERROR:  no relation entry for relid 2 (relnode.c:199)
Now for 'view update', result_relation is a view relation and
Var created based on result_relation will break the functionality.

In this patch we also remove the code that adds the "ctid" entry
to targetList in preprocess_targetlist() since 1) upstream
removes this also in that function 2) Looks like
a new function rewriteTargetListUD() adds that for regular
table at the rewrite stage.

I added a FIXME since I'm not sure how to handle for other relation
for column gp_segment_id at this moment.