Fix broken extract_actual_join_clauses call in 9.6 postgres_fdw.
In commits e5d83995e et al, I changed the signature of
extract_actual_join_clauses, thinking that it was not called from
anywhere but createplan.c.  I missed that postgres_fdw uses it
in the 9.6 branch only.

This opens up the question of whether any third-party modules might
be calling it, and whether we need to take steps to avoid an API break
for them.  But for the moment, just get the buildfarm green again.