Address parent ResultRelInfo correctly when getting partition.
ResultRelInfos for partitions are now kept as part of the parent
ResultRelInfo, not in a global hash table (estate->es_partition_state->
result_partition_hash) any more. This is changed by commit a7daa7.

When getting a partition, the parent ResultRelInfo is not addressed
correctly, that leads to a ResultRelInfo to add to itself.

This commit fixes that and would hopefully fix 'tuple already updated by
self' in gpcheckcat.