Use repeatable read when serializable isolation is requested
This was the behavior in upstream, prior to 9.1.  Greenplum needs a way
to detect serialization conflicts by reconciling predicate locks from
all segments in the cluster.  Until that is implemented, this temporary
patch makes serializable fall back to repeatable read.  A fixme is added
so that we remember to implement true serializability.

Only a log message is generated when falling back.  A warning / error
may break compatibility with existing applications.  If needed, the log
message may be turned into a warning (and many expected output files
updated in regression tests) later.