Use default collation during ORCA DXL to PlStmt translation
This is a follow-on to the commit 72e75e31, that determines the
collation to use for the various plan nodes in the translator based on
the type of the element.

If the type is collatable, we use DEFAULT_COLLATION_OID, otherwise we
use InvalidOid.

NOTE: It should also probably be squashed with the previous commit
72e75e31.

Signed-off-by: Melanie Plageman <melanieplageman@pivotal.io>