Disable INSTEAD OF triggers temporarily
INSTEAD OF triggers are new as of PostgreSQL 9.1.  They allow
INSERT/UPDATE/DELETE operations on views.  Greenplum is not ready to
plan DML on views.  ModifyTable node treats a view as master-only
relation and the node never gets dispatched to QEs.  In case the DML
contains a where clause and direct dispatch becomes possible, this
results in error during commit: "Distributed transaction ... not
found".  Disable the feature in Greenplum and ignore the diffs for
tests that came with it until the plan generation for DML on views is
fixed.